Create getWordPressVersion to be used in CLI and Studio electron app
